Base Version (Does not contain all Commercial Features)
This Version of CDN in a Box doesn't do all the tricks of the full
version, but offers a lot of features for Bloggers, and Content sites.

Post requests work only in HTTP, and Cookies aren't fully supported in
this version so it may not be suitable for Web Apps.